+++ Sources/SwiftFortuneWheel/Utils/Drawing/SliceDrawing.swift
@@ -35,6 +35,11 @@ extension SliceDrawing {
         self.preferences?.contentMargins.top ?? 0
     }
 
+    /// Content bottom margin
+    var bottomMargin: CGFloat {
+        self.preferences?.contentMargins.top ?? 0
+    }
+
 
     /// Context position correction offset degree
     var contextPositionCorrectionOffsetDegree: CGFloat {
@@ -75,13 +80,8 @@ extension SliceDrawing {
         // Draws slice content
         slice.contents.enumerated().forEach { (contentIndex, element) in
             switch element {
-            case .text(let text, let preferenes):
-                if preferenes.isCurved {
-                    self.drawCurved(text: text, in: context, preferences: preferenes, rotation: rotation, index: index, topOffset: topOffset)
-                } else {
-                    self.draw(text: text, in: context, preferences: preferenes, rotation: rotation, index: index, topOffset: topOffset)
-                }
-                topOffset += preferenes.preferedFontSize + preferenes.verticalOffset
+            case .text(let text, let preferences):
+                topOffset += prepareDraw(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
             case .image(let name, let preferenes):
                 self.drawImage(in: context, imageName: name, preferences: preferenes, rotation: rotation, index: index, topOffset: topOffset)
                 topOffset += preferenes.preferredSize.height + preferenes.verticalOffset
@@ -176,6 +176,31 @@ extension SliceDrawing {
         context.restoreGState()
     }
 
+    /// Prepare to draw text
+    /// - Parameters:
+    ///   - text: text
+    ///   - context: context where to draw
+    ///   - preferences: text preferences
+    ///   - rotation: rotation degree
+    ///   - index: index
+    ///   - topOffset: top offset
+    /// - Returns: height of the drawn text
+    func prepareDraw(text: String, in context: CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) -> CGFloat {
+        switch preferences.orientation {
+        case .horizontal:
+            if preferences.isCurved {
+                self.drawCurved(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+            } else {
+                self.drawHorizontal(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+            }
+        case .vertical:
+            // Not finished yet
+//            self.drawVertical(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+            self.drawHorizontal(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+        }
+        return preferences.preferedFontSize + preferences.verticalOffset
+    }
+
     /// Draws curved text
     /// - Parameters:
     ///   - text: text
@@ -218,7 +243,7 @@ extension SliceDrawing {
     ///   - rotation: rotation degree
     ///   - index: index
     ///   - topOffset: top offset
-    private func draw(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
+    private func drawHorizontal(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
 
         var textColor: UIColor = .black
 
@@ -247,11 +272,77 @@ extension SliceDrawing {
         context.saveGState()
         context.rotate(by: rotation * CGFloat.pi/180)
 
+        if preferences.flipUpsideDown {
+            context.rotate(by: flipRotation)
+        }
+
         text.draw(in: CGRect(x: -(textRect.width / 2) - preferences.horizontalOffset, y: yPosition, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
+
+        context.restoreGState()
+
+    }
+
+    /// Draws text
+    /// - Parameters:
+    ///   - text: text
+    ///   - context: context where to draw
+    ///   - preferences: text preferences
+    ///   - rotation: rotation degree
+    ///   - index: index
+    ///   - topOffset: top offset
+    private func drawVertical(text: String, in context: CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
+
+        var textColor: UIColor = .black
+
+        switch preferences.textColorType {
+        case .evenOddColors(let evenColor, let oddColor):
+            textColor = index % 2 == 0 ? evenColor : oddColor
+        case .customPatternColors(let colors, let defaultColor):
+            textColor = colors?[index, default: defaultColor] ?? defaultColor
+        }
+
+        let textFontAttributes: [NSAttributedString.Key:Any] = {
+            let textStyle = NSMutableParagraphStyle()
+            textStyle.alignment = .center
+            let deafultAttributes:[NSAttributedString.Key: Any] =
+                [.font: preferences.font,
+                 .foregroundColor: textColor,
+                 .paragraphStyle: textStyle ]
+            return deafultAttributes
+        }()
+
+        let textWidth: CGFloat = text.width(by: preferences.font)
+
+        let maxAvailableBottomRadiusOffsetCircularSegmentHeight = self.leftMargin + self.rightMargin + preferences.preferedFontSize
+        let bottomRadiusOffset = max(self.bottomMargin, radius(circularSegmentHeight: maxAvailableBottomRadiusOffsetCircularSegmentHeight, from: sliceDegree))
+
+        let availableHeightInSlice: CGFloat = radius - preferences.verticalOffset - topOffset - bottomRadiusOffset
+        let textRectWidth = min(textWidth, availableHeightInSlice)
+        let croppedText = text.crop(by: textRectWidth, font: preferences.font)
+
+//        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.preferedFontSize)
+        let textRect = CGRect(x: 0, y: 0, width: textRectWidth, height: preferences.preferedFontSize)
+        let yPosition = -(radius - preferences.verticalOffset - topOffset - topMargin)
+        let xPos = -(textRect.width / 2) - preferences.horizontalOffset
+
+        context.saveGState()
+        context.rotate(by: rotation * CGFloat.pi/180)
+        context.translateBy(x: xPos + textRect.height / 2, y: yPosition + textRect.width / 2)
+        context.rotate(by: 90 * CGFloat.pi/180)
+
+
+//        context.rotate(by: 90 * CGFloat.pi/180)
 
         if preferences.flipUpsideDown {
             context.rotate(by: flipRotation)
         }
+
+//        croppedText.draw(in: CGRect(x: 0, y: 0, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
+//        croppedText.draw(in: CGRect(x: -(textRect.width / 2) - preferences.horizontalOffset - rotationOffset / 2, y: yPosition - rotationOffset / 2 , width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
+//        context.addRect(textRect)
+//        UIColor.red.setStroke()
+//        context.drawPath(using: .fillStroke)
+        croppedText.draw(in: CGRect(x: -textRect.width / 2, y: -textRect.height / 2, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
 
         context.restoreGState()